renderers/governor/requirements.txt
python-dateutil
GitPython
requests